Move users controller to root move login function to app controller
