+++ src/users.controller.ts
@@ -1,0 +1,11 @@
+import { Controller, UseGuards, Request, Get } from '@nestjs/common';
+import { JwtAuthGuard } from './auth/jwt-auth.guard';
+
+@Controller('users')
+export class UsersController {
+  @UseGuards(JwtAuthGuard)
+  @Get('profile')
+  getProfile(@Request() req) {
+    return req.user;
+  }
+}
